SAMZA-2014: Samza-sql: Support table as both source (for join) and destination in the same application
While parsing queries in an application, with in SamzaSqlApplicationConfig, we collect all input sources and output sources from all queries and create descriptors for input sources first followed by output sources. But there could be only one table descriptor instance per table. Writable table is a readable table but vice versa is not true. If we go through input sources, we will end up creating readable table descriptor and would not be able to create writable table descriptor again when we go through output sources (the code will be ugly if we have to achieve this). There are couple of ways to solve this:
- Always make a table readable and writable
- Go through output sources first followed by input sources.

Choosing option 2 as making a table always read-writable does not make sense.

Author: Aditya Toomula <atoomula@linkedin.com>

Reviewers: Srinivasulu Punuru <spunuru@linkedin.com>

Closes #834 from atoomula/table