SAMZA-2022: Integrate startpoints with SystemConsumers
This PR integrates the `Startpoint` with the `SystemConsumers` multiplexer class. The general logic is if there is a `Startpoint` for an SSP in a task and the `SystemConsumer` implements the `StartpointVisitor` interface, it will apply the visitor method via `Startpoint#apply(SystemStreamPartition, StartpointVisitor)`. Otherwise, the `SystemConsumer#register(SystemStreamPartition, checkpointOffset:String)` method is called as it currently does.

shanthoosh cameronlee314 prateekm - Please take a look... after the break is fine. If I don't respond to comments within 24-48 hours during my leave, please feel free to hijack this PR into another PR.

Author: Daniel Nishimura <dnishimura@linkedin.com>

Reviewers: Shanthoosh Venkataraman <spvenkat@usc.edu>,  Jake Maes<jmaes@linkedin.com>, Cameron L<calee@linkedin.com>

Closes #869 from dnishimura/samza-2022-startpoint-systemconsumers-integ