SAMZA-2018: Using Rocksdb bulk load only for clean-slate restores
Rocksdb has a problem that its bulk-load option cannot be used on existing stores seamlessly.
See https://github.com/facebook/rocksdb/issues/2734

Because of this when host-affinity puts a container onto an old host, the restore path's bulk-load mode fails.

To fix this,
We use rocksdb's bulk load option only when the underlying storeDirectory does not exist.
If the storeDirectory exists, then we open the store without bulk-load options.

Note that, an existing storeDirectory may also be "stale", in which case we delete it, and should ideally be using the bulk load option. We will add this additional optimization separately.

Author: Ray Matharu <rmatharu@linkedin.com>
Author: Jacob Maes <jacob.maes@gmail.com>

Reviewers: Bharath Kumarasubramanian <bkumaras@linkedin.com>

Closes #892 from rmatharu/rocksdbbugfix