SAMZA-2117: Handle race condition in container launch due to incorrect AM accounting
Steps involved when starting a Samza container:
1. Issue a request to YARN to launch a container
2. Record that container as "pending" launch.
3. Launch callback succeeds on a different thread : The callback looks at the "pending" container and marks it as "running".

A race-condition in the above: If the main thread gets pre-empted between (1) and (2), the callback thread wouldn't see the container state as "pending" - hence, it wouldn't transition it to a "running" state.

This PR fixes it by flipping (1) and (2) - ie., record the intent prior to issuing the launch request.
Added an unit test - refactored existing tests

Author: Jagadish <jvenkatraman@linkedin.com>

Reviewers: Prateek M<pmaheshw@linkedin.com>

Closes #958 from vjagadish1989/samza-2117