SAMZA-2170: Enabling writing of both new and old format offset files for stores and side-input-stores
After Samza 1.1, the offset file for stores and sideinputs has been unified and is versioned.
However, this Jira adds the logic in code to read and write both this new and old format. Because of this apps can switch between 1.0 and 1.1 versions seamlessly.

Note that the old format and filenames for store and side-input offset differed.

Author: Ray Matharu <rmatharu@linkedin.com>

Reviewers: Prateek Maheshwari <pmaheshwari@apache.org>

Closes #1005 from rmatharu/bugfix-offset