SAMZA-2188: Remove dependency to com.couchbase.client:encryption
com.couchbase.client:encryption is a module that is needed on client side when client side encryption is enabled. It is unfortunately a commercial lib which is not free. Therefore we are excluding it by default, end user could manually include it when required.

More info can be found here

https://mvnrepository.com/artifact/com.couchbase.client/encryption

Author: Wei Song <wsong@linkedin.com>

Reviewers: Srini Pururu <spunuru@linkedin.com>

Closes #1024 from weisong44/SAMZA-2188