SAMZA-2274: Cleanup store directories on startup for non-logged store only
During startup, the CSM cleans up the base directories for stores. Ideally, this cleanup should only be done for non-logged stores. In our current setup, if the applications configure both logged and non-logged store base directories to be the same, we delete the logged store directories as well which results in unnecessary bootstrap.

In this PR,
 * We modify the clean up behavior on startup to impact only non-logged stores
 * Update the configuration documentation to reflect potential implications of not configuring `job.logged.store.base.dir` and `job.non.logged.store.base.dir` for stateful applications.
 * Add a warning if both logged and non-logged store base directories are configured to use same path.

Author: mynameborat <bharath.kumarasubramanian@gmail.com>

Reviewers: Hai Lu <halu@linkedin.com>

Closes #1107 from mynameborat/SAMZA-2274 and squashes the following commits:

7d47a90b [mynameborat] Fix the logged store condition check
35b1bb3d [mynameborat] Cleanup store directories on startup for non-logged store only