Transactional State [1/5]: Added Store Checkpoint API
This PR adds a new 'checkpoint' API to KeyValueStore that allows taking persistent (on-disk) snapshots of the local store state during commits for implementations that support it.

In RocksDB, this creates a new Checkpoint directory.

For supporting transactional state, during a commit, the newest changelog SSP offset is stored in the OFFSET file in this checkpoint directory, as well as in the checkpoint topic. During container start / store restore, the checkpoint directory corresponding to the changelog offset in the checkpoint topic is selected (if any) and used as the basis for restore.

Thanks to @xinyuiscool for contributing this.