Transactional State [2/5]: Added a ChangelogSSPIterator API
This PR changes the KeyValueStorageEngine restore API and implementation to take a new ChangelogSSPIterator instead of a plain Iterator.

The new ChangelogSSPIterator is similar to the existing SystemStreamPartitionIterator with the following differences:

This new iterator has 2 modes: Restore and Trim.
It takes an endingOffset during construction. The mode changes from Restore to Trim during iteration if the current message offset is greater than the ending offset (and trim mode is enabled).
Does not check for end of stream since it isn't applicable to changelog topics.
For supporting transactional state, we only restore changelog messages up to the changelog SSP offset in the checkpoint topic. Any messages after this 'checkpointed changelog offset' are trimmed by overwriting them with the current store value. When used in conjunction with an appropriate 'min.compaction.lag.ms' configuration for the Kafka changelog topic, this ensures that on container restart any store contents are consistent with the last input checkpoints and do not reflect any newer changes.