SAMZA-2562: [Scala cleanup] Clean up Scala in samza-kv-inmemory and samza-kv-couchbase (#1404)
API changes: I don't think the InMemoryKeyValueStore is part of the Samza API, since InMemoryKeyValueStorageEngineFactory is supposed to be used, but the InMemoryKeyValueStore constructor did change from having one KeyValueStoreMetrics argument with a default value of new KeyValueStoreMetrics to having a required KeyValueStoreMetrics argument. I believe this only applies to Scala usage of InMemoryKeyValueStore, since Java usage requires all arguments (even ones with default values) to be passed anyways.

Upgrade/usage instructions: If using the InMemoryKeyValueStore constructor in another Scala class, then add the KeyValueStoreMetrics argument. It's unlikely that anyone is depending on this constructor explicitly.