SAMZA-2593: Update task callback to store only necessary fields instead of the message envelope (#1433)
Problem: TaskCallbackImpl currently stores the entire message envelope. In case of jobs that have 1000s of pending callbacks, this adds up to significant memory pressure. The only use for message envelope is to trace the SSP information and the offset associated with it so that we commit the right offset during our commit cycle.

Changes: Modify TaskCallbackImpl to only store SSP and Offset information.