SAMZA-2618: Add per-container metric for memory-utilization (#1459)
Feature: Add a metric to measure the fraction of physical memory used out of the total memory of the container. During autosizing, the container-count and size can change as traffic changes, so only monitoring the phsical-memory-mb does not provide enough information for alerting. We need a fraction or % metric to allow for easy monitoring and alerting

Changes: Add a metric to measure the fraction of physical memory used out of the total memory of the container.