SAMZA-2621: Support custom operator name in MessageStream
Feature: Add a new set of methods in MessageStreamm, which takes a customize operator name. This operator name will be composed as part of operator id when not empty. Therefore, metrics, logs will include the customize name in it. The purpose of this change is to provide easier approach for clients to understand their metrics/logs.
Changes: For each method in MessageStream, a new method is provided with an extra description parameter. Existing methods will defaults to the new methods with empty description.
Tests: Unit Tests
API Changes: A new set of operator methods in MessageStream.
Upgrade Instructions: N/A since API changes are backward compatible.
Usage Instructions:

For stateless operations such as map, flatMap, flatMapAsync, filter, sink, sendTo and merge, user may simply supply an addition param to name their operators. e.g.

map($func) -> map("description of the map operation", $func)

This results in a change of operator id, therefore, a change in metrics reported.

For stateful operations such as window, join, partitionBy and broadcast, supplying an customized operation name is backward incompatible as change in operator id changes the default store name. This is not recommended.