SAMZA-2632: Processors should only write their own task locality to zookeeper (#1473)
Problem:
Processors update the task locality for all the tasks in the job model with their own locality regardless of whether the tasks belong to their container model or not.
Description:
As part of SEP 11, host affinity for standalone was introduced. For this feature, we persist the task locality in zookeeper so that subsequent rebalances take this locality into account when generating job model. During job model consensus, we update the task locality for all the tasks and this results in incorrect locality depending on the order of writes.
Changes:
Only update the locality for the tasks that belong to the processor.