Changed StreamOperatorTask to throw appropriate Exception when InputOperator is not found for a SystemStream in the OperatorGraph (#1483)
Symptom:
Samza job fails with a TaskCallbackTimeoutException if InputOperator is not found in OperatorGraph for a SystemStream

Cause:
Currently, the pipeline silently fails when there is no InputOperator for a SystemStream in the OperatorGraph as the code doesn't handle such cases. The pipeline just silently fails and the job would throw an exception due to sync task callback timeout (TaskCallbackTimeoutException).
One of the scenarios that causes this is due to non-determinism in the plan seen by the containers and the AM. It can happen in multiple scenarios where the transform/operators have non-determinism in generating their names. e.g. we noticed beam transform appended hashcode.

Changes:
Call failure method of the TaskCallback with a SamzaException with the appropriate exception message.