SAMZA-2633: Rolling upgrades cause downtime to upgraded processors for the entire deployment window (#1484)
Description:
During rolling upgrades, the current debounce timer gets extended every time when there is a quorum change notification. As a result, processors that were upgraded earlier in the deployment window remain unavailable waiting for work assignment. In some scenarios, this cause processors to be unavailable for 20 minutes or so depending on the size of the quorum and the debounce time configuration. Refer to SAMZA-2633 for more information.

Changes:
Optimize the leader workflow to skip rebalance if there is no changes to work assignment
Make processors start with most recent agreed job model on startup
Leader persists the active job model version in ZK to enable change [2]
Introduce config for applications to opt-in for the optimization

Usage Instructions:
Set job.coordinator.zk.enable-startup-with-active-job-model to true as part of the application configuration to enable processor use the recent active job model during startup and also enable leader to skip rebalances if the work assignment remains the same.