SAMZA-2649: Add MetricsReporter which logs metrics to log file (#1495)
Issues: In environments without an external metrics system, it can sometimes be hard to access any metrics from a Samza job. It can be useful to have a MetricsReporter which logs metrics so that they can be accessed in a simple way with minimal dependencies. This can be used to help verify certain flows are working. An example case for using this is when testing Samza on a bare-bones Kubernetes cluster in minikube.

Changes: Added LoggingMetricsReporter which periodically logs metrics to a log file.

API changes and usage/upgrade instructions:
Add a new metrics reporter in the config which uses org.apache.samza.metrics.reporter.LoggingMetricsReporterFactory. See https://github.com/apache/samza/pull/1495 for an example of configuration.