SAMZA-2669: Inject MetricsRegistry into SamzaContainer so that it can be shared in higher layers (#1518)
Changes:
1. Inject a MetricsRegistryMap as an argument into SamzaContainer.apply
2. Remove unused name field from MetricsRegistryMap.

API changes and upgrade/usage instructions: MetricsRegistryMap no longer has a "name" associated with it. There is no longer a constructor for MetricsRegistryMap which accepts a name, and MetricsRegistryMap no longer has a getName method. Since MetricsRegistryMap is in samza-core, this is not a public API, but putting a note here just in case.