SAMZA-2685: Add job coordinator which does not do resource management (#1529)
Adding a job coordinator which does not do resource management. For Samza on YARN, the ClusterBasedJobCoordinator does resource management. However, for Samza on Kubernetes, it is not necessary to have a job coordinator which does resource management, because Kubernetes controllers can take care of resource management.

API changes (all changes are backwards compatible):
1. Set the config job.coordinator.factory to org.apache.samza.coordinator.staticresource.StaticResourceJobCoordinatorFactory in order to use the new coordinator.
2. Set the config job.coordinator.restart.signal.factory to define how to restart the Samza job when an input stream changes which will change the job model. This plug-in is dependent on where the Samza job is running (e.g. Kubernetes). Currently, there is only a no-op implementation of this restart signal.