SAMZA-2075: Add execEnvContainerId and execEnvAttemptId for diagnostics for Kubernetes job coordinator (#1566)
API changes (backwards compatible):
JSON representation of MetricsHeader (used in MetricsSnapshot) has an additional optional field called samza-epoch-id, and this field is intended to contain the deployment attempt id which is shared across all containers (job coordinator and wokrers) of a Samza job. This field is filled in by reading the SAMZA_EPOCH_ID environment variable.