Emit to diagnostic kafka topic in round robin fashion to support increasing this topic partition count (#1631)
Symptom: For some Samza jobs, the size of the diagnostic stream partition grows to be large and can cause issues in system streams like Kafka

Cause:
Currently, diagnostics topics for Samza jobs are created with a single partition.
Because it is created in DiagnosticsUtil and uses the default partition count, which is 1.
There could be a need to increase the partition count of the diagnostic stream to resolve the issue above.
However, Samza currently uses hostnames as partition key to distribute diagnostic msgs. The host names can be very similar, the hash of these in the partition key will not evenly distribute the msgs across partitions.

Change:
Use the “null” partition key instead of using the hostname as the partition key, and leverage on Kafka default partitioner to evenly distributed the events across partitions.
Note that this change is backward compatible with the default partition count (1) since every msg goes to the same partition regardless of the partition key.

API changes: None

Backward compatible: No. Note that the change will be backwards incompatible only when the diagnostic stream has more than one partition. The default creation for diagnostic stream inside samza is with 1 partition, but there are cases that the stream pre-exists or partition count increased after creation.

Test Done:
./gradlew build