Fix deadlock in StreamAppender (#1652)
Issue:
In StreamAppender, there is a synchronized block around setupSystem(), which should be called the first time (after loggingContext config is set up) any logger logs something.
It could lead to a deadlock situation. Because if during the setupSystem(), any other threads try to do LOG.xxx(), they will be blocked, and if the system setup depends on those threads, it leads to deadlock

Change:
Replace the synchronized block with a tryLock(long time, TimeUnit unit)
so that other threads won't be blocked forever if setupSystem() stuck or timeout

Usage Change:
If any thread try to log during the time that another thread is setting up the stream, it will wait for SET_UP_SYSTEM_TIMEOUT_MILLI_SECONDS. If timeout, StreamAppender will skip sending its log event

Note that before the change, the thread will block and wait until the stream setting up is done. This behavior opens the chance of deadlock and needs to be changed