SAMZA-2795: Set thread to daemon thread for operator executor service (#1690)
Description
As part of SAMZA-2781, we introduced operator executors to manage operator handoff execution. However, the threads created by the executor service are non-daemon and hence prevent the JVM from shutting down.

For context, we don't have a clean way to shutdown the executor due to lack of clean lifecycle management of the factory. Hence shutting down the executor service within TaskInstance is not an option and the fix is to make it daemon threads.

Changes
Make the threads spawned by the operator executor to be non-daemon

Tests
None

API Changes
None

Upgrade Instructions
None

Usage Instructions
None